Move bulk of webhook logic to plugin
