+++ plugins/webhooks/client/Icon.tsx
@@ -1,0 +1,1 @@
+export { WebhooksIcon as default } from "outline-icons";
