+++ plugins/webhooks/plugin.json
@@ -1,0 +1,5 @@
+{
+  "name": "Webhooks",
+  "description": "Adds HTTP webhooks for various events.",
+  "requiredEnvVars": []
+}
