+++ plugins/webhooks/server/.babelrc
@@ -1,0 +1,3 @@
+{ 
+  "extends": "../../../server/.babelrc"
+}
